move loginHelper to parent utils
