--- expressEngine/login/utils/index.js
@@ -1,1 +1,0 @@
-export { default } from './login';
